move AwsPrivateKeyStore from server_common to auth_core as AwsS3PrivateKeyStore
